Refactored triplestore module
