+++ docs/api_reference/errors.md
@@ -1,0 +1,3 @@
+# errors
+
+::: tripper.errors
