+++ docs/api_reference/literal.md
@@ -1,0 +1,3 @@
+# literal
+
+::: tripper.literal
