+++ docs/api_reference/namespace.md
@@ -1,0 +1,3 @@
+# namespace
+
+::: tripper.namespace
