+++ tripper/backends/rdflib.py
@@ -11,7 +11,7 @@ from rdflib import BNode, Graph
 from rdflib import Literal as rdflibLiteral
 from rdflib import URIRef
 
-from tripper.triplestore import Literal
+from tripper import Literal
 
 if TYPE_CHECKING:  # pragma: no cover
     from collections.abc import Sequence
